feat(ui): move strongly-related utils
